add astype

+++ sdgx/metrics/pair_column/mi_sim.py
@@ -48,10 +48,10 @@ class MISim(PairMetric):
 
         if data_type == "numerical":
             src_col = pd.cut(
-                src_col, instance.numerical_bins, labels=range(instance.numerical_bins)
+                src_col.astype(float), instance.numerical_bins, labels=range(instance.numerical_bins)
             )
             tar_col = pd.cut(
-                tar_col, instance.numerical_bins, labels=range(instance.numerical_bins)
+                tar_col.astype(float), instance.numerical_bins, labels=range(instance.numerical_bins)
             )
             src_col = src_col.to_numpy()
             tar_col = tar_col.to_numpy()
